+++ e2e/next-feeds/app/api/auth/liveblocks/route.ts
@@ -13,7 +13,7 @@ const USERS = [
 
 /**
  * Authenticating your Liveblocks application
- * https://liveblocks.io/docs/authentication
+ * https://liveblocks.io/docs/api-reference/authentication
  */
 
 const liveblocks = new Liveblocks({