+++ examples/nextjs-nextauth-google-avatars/pages/api/liveblocks-auth.ts
@@ -5,7 +5,7 @@ import { User } from "../../types";
 
 /**
  * Authenticating your Liveblocks application
- * https://liveblocks.io/docs/authentication
+ * https://liveblocks.io/docs/api-reference/authentication
  */
 
 const liveblocks = new Liveblocks({