+++ examples/nextjs-notifications-custom/src/app/api/liveblocks-auth/route.ts
@@ -4,7 +4,7 @@ import { getUser } from "../../../database";
 import { getExampleUserId } from "../../../example";
 
 // Authenticating your Liveblocks application
-// https://liveblocks.io/docs/authentication
+// https://liveblocks.io/docs/api-reference/authentication
 
 const liveblocks = new Liveblocks({
   secret: process.env.LIVEBLOCKS_SECRET_KEY as string,