+++ examples/nextjs-tiptap/app/api/liveblocks-auth/route.ts
@@ -4,7 +4,7 @@ import { getRandomUser } from "../database";
 
 /**
  * Authenticating your Liveblocks application
- * https://liveblocks.io/docs/authentication
+ * https://liveblocks.io/docs/api-reference/authentication
  */
 
 const liveblocks = new Liveblocks({