+++ examples/nextjs-yjs-slate/src/app/api/liveblocks-auth/route.ts
@@ -3,7 +3,7 @@ import { NextRequest } from "next/server";
 
 /**
  * Authenticating your Liveblocks application
- * https://liveblocks.io/docs/authentication
+ * https://liveblocks.io/docs/api-reference/authentication
  */
 
 const liveblocks = new Liveblocks({