+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## vNEXT (not yet released)
 
+### `@liveblocks/node`
+
+- Add a `hasThreads` query filter to `getRooms()` and `iterRooms()`.
+
 ### `@liveblocks/chat-sdk-adapter`
 
 - Allow `apiKey` and `webhookSecret` to be resolved per request, and add custom